# Optimize vis_heap_chunks command