--- a/profiling/benchmark_vis_heap_chunks/README.md
+++ b/profiling/benchmark_vis_heap_chunks/README.md
@@ -1,0 +1,2 @@
+This benchmark was used to investigate performance problems with the `vis_heap_chunks` command described in https://github.com/pwndbg/pwndbg/issues/1675
+