--- a/profiling/benchmark_vis_heap_chunks/bench.sh
+++ b/profiling/benchmark_vis_heap_chunks/bench.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-gdb --batch --ex 'break exit' --ex 'run' --ex 'source gdbscript.py' --args `which python3` -c 'import sys; sys.exit(0)'
+gdb --batch --ex 'break exit' --ex 'run' --ex 'source gdbscript.py' --args $(which python3) -c 'import sys; sys.exit(0)'