--- a/profiling/benchmark_vis_heap_chunks/bench.sh
+++ b/profiling/benchmark_vis_heap_chunks/bench.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+gdb --batch --ex 'break exit' --ex 'run' --ex 'source gdbscript.py' --args `which python3` -c 'import sys; sys.exit(0)'