--- a/pwndbg/commands/heap.py
+++ b/pwndbg/commands/heap.py
@@ -3,6 +3,7 @@ import ctypes
 
 import gdb
 from tabulate import tabulate
+from typing import Dict, List
 
 import pwndbg.color.context as C
 import pwndbg.color.memory as M